Updated connection pool,dao and entities

+++ FinalProject/src/main/java/com/epam/db/dao/PaymentDao.java
@@ -8,7 +8,7 @@ import com.epam.exceptions.DaoException;
 import java.util.List;
 import java.util.Optional;
 
-public interface PaymentDao extends BaseDao<Long,Payment>{
+public interface PaymentDao extends BaseDao<Integer,Payment>{
     Optional<Payment> findPaymentByCriteria(PaymentCriteria paymentCriteria) throws DaoException;
 
     List<Payment> findAllPaymentByCriteria(PaymentCriteria paymentCriteria) throws DaoException;
