+++ FinalProject/src/main/java/com/epam/db/dao/RequestDao.java
@@ -7,7 +7,7 @@ import com.epam.exceptions.DaoException;
 import java.util.List;
 import java.util.Optional;
 
-public interface RequestDao extends BaseDao<Long,Request>{
+public interface RequestDao extends BaseDao<Integer,Request>{
     Optional<Request> findRequestByCriteria(RequestCriteria requestCriteria) throws DaoException;
 
     List<Request> findAllRequestByCriteria(RequestCriteria requestCriteria) throws DaoException;
